implementation of Product Service

+++ product-service/src/main/java/managment/productservice/controller/request/ProductCreateRequest.java
@@ -10,12 +10,12 @@ import java.math.BigDecimal;
 @Data
 @Setter
 @Getter
-// @ApiModel(value = "Product Create Request", description = "Product create request")
+// @ApiModel(value = "Product Create & Update Request", description = "Product create request")
 public class ProductCreateRequest {
     private String productName;
     private String productDescription;
     private String productCategory;
     private double productPrice;
-    private BigDecimal productDiscount;
+    private double productDiscount;
 
 }
